add option to use an existing AudioContext
Since Hark creates an audio context each time it is initialized, it is
not a good option to monitor audio levels or speech events in situations
where there are multiple streams that require it. Only a single
`AudioContext` is required per `document`, so it would be advantageous to
allow passing an existing one to Hark, if required.